requirements.txt
# fastapi==0.104.1
# uvicorn==0.24.0
# pydantic==2.5.0
# openai==1.3.0
# python-dotenv==1.0.0
# pandas==2.1.3
# numpy==1.25.2
# scikit-learn==1.3.2
# requests==2.31.0
# python-multipart==0.0.6

fastapi==0.104.1
uvicorn==0.24.0
pydantic>=2.9.0
openai==1.3.0
python-dotenv==1.0.0
pandas>=2.2.3
numpy>=1.26.4
scikit-learn>=1.5.0
requests==2.31.0
python-multipart==0.0.6
